discord_bot_project/requirements.txt
discord.py==2.3.0
aiohttp==3.8.4
requests==2.28.1
flask==2.3.2
psycopg2==2.9.5
pytz==2023.3
python-dotenv==0.21.1
